requirements.txt
annotated-types==0.6.0
anyio==3.7.1
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.7
dnspython==2.4.2
email-validator==2.1.0.post1
fastapi==0.104.1
greenlet==3.0.2
h11==0.14.0
idna==3.6
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
PyMySQL==1.1.0
python-dotenv==1.0.0
pytz==2023.3.post1
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.24.0.post1
